fix: can name a test file named after an binary
If one named its test file after an binary that can be found in
PATH then bash_unit was trying to source this binary instead of
the test file.

Also note that the path displayed in test output may not be the
exact path of the file as bash_unit changes directory to the test
file directory before running the test so it's a local path that is
displayed. That's a problem I keep for later.